Make withRecovery rewind before recovering
This makes it easier to control the fast-forwarding you often want to do
when using withRecovery, because you have more control over where it
starts.

For example, if you previously did `many (withRecovery runToNextNewline
p)`, and `p` fails immediately at the start of a newline, you ran the
risk of looping unless `runToNextNewline` was implemented as always
consuming at least one character. The problem with doing that, prior to
this change, was that if `p` failed at the end of a line, the recovery
would skip not to the next newline, but to the one after.

With this change, you can implement `runToNextNewline` as always
consuming at least one character, and rest assured that it'll always
skip precisely to the next newline.

As far as I can tell, this deviates slightly from what Megaparsec does,
but seems better to me, though possibly a bit less performant.